requirements.txt
werkzeug
Flask
Flask-SQLAlchemy
Flask-WTF
flask-bootstrap
flask-paranoid
marshmallow-sqlalchemy
flask_marshmallow
flask-login
marshmallow
mysqlclient
python-dotenv